requirements.txt
fastapi
uvicorn
sqlalchemy
passlib[bcrypt]
python-jose
requests
pydantic
bcrypt==4.0.1
